refactor llms code using OOP, and model registry, you need to inherit baseLLM and registry model when you want to add a new model
