+++ src/llms/llms_v2/constant.py
@@ -1,0 +1,5 @@
+from transformers import AutoModelForCausalLM
+
+MODEL_CLASS = {
+    "causal_lm": AutoModelForCausalLM,
+}
